requirements.txt
ase==3.22.1
datasets==2.18.0
fschat==0.2.36
hydra-core==1.3.2
ipython==8.23.0
matplotlib==3.8.4
numpy==1.26.4
torch==2.1.2
omegaconf==2.3.0
peft==0.10.0
pytorch_lightning==2.2.2
sentencepiece==0.2.0
setproctitle==1.3.3
torchrl==0.2.1
torch_optimizer==0.3.0
tqdm==4.66.2
transformers==4.39.3
wandb==0.16.6
hydra-submitit-launcher==1.2.0
openai==1.30.5